requirements.txt
Django==5.0.2
djangorestframework==3.14.0
PyQt5==5.15.9
psutil==5.9.5
pynput==1.7.6
pygetwindow==0.0.9
requests==2.31.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
django-redis==5.4.0
django-cors-headers==4.3.1 